Add option to continue after unknown result
Currently, the trace executor only asks the solver to solve only
`(check-sat)` commands up to the first command with an unknown status.
This commit adds the option `--continue-after-unknown` which makes the
trace executor continue after an expected `unknown` result. This option
is useful for determining previously unknown statuses of benchmarks.